# docs/swagger.yaml
basePath: /
definitions:
  dto.AssignRoleRequest:
    properties:
      roleId:
        type: string
    required:
    - roleId
    type: object
...
  dto.CreateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.PermissionResponse:
    properties:
      canCreate:
        type: boolean
      canDelete:
        type: boolean
      canRead:
        type: boolean
      canUpdate:
        type: boolean
      permissions:
        type: integer
      resource:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResponse'
        type: array
    type: object
...
  dto.SetPermissionRequest:
    properties:
      permissions:
        maximum: 15
        minimum: 0
        type: integer
      resource:
        maxLength: 50
        type: string
    required:
    - resource
    type: object
...
  dto.UpdateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    type: object
...
  dto.UserPermissionResponse:
    properties:
      canCreate:
        type: boolean
      canDelete:
        type: boolean
      canRead:
        type: boolean
      canUpdate:
        type: boolean
      permissions:
        type: integer
      resource:
        type: string
      userId:
        type: string
    type: object
  dto.UserRolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      userId:
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Clean Architecture Go API with GORM, Gin, PostgreSQL, Redis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/permissions:
    get:
      consumes:
      - application/json
      description: Get the current user's permission on a specific resource
      parameters:
      - description: Resource name (e.g., blogs, categories)
        in: query
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserPermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get my permission on a resource
      tags:
      - Roles
...
  /api/v1/roles:
    get:
      consumes:
      - application/json
      description: List all available roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role (admin only)
      parameters:
      - description: Create role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a new role
      tags:
      - Roles
  /api/v1/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role (admin only)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get a role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update a role (admin only)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a role
      tags:
      - Roles
  /api/v1/roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Set permission for a role on a specific resource (admin only)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Set permission request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Set permission for a role
      tags:
      - Roles
...
  /api/v1/users/{userId}/roles:
    get:
      consumes:
      - application/json
      description: Get all roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserRolesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get user roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Assign a role to a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Assign role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Assign role to user
      tags:
      - Roles
  /api/v1/users/{userId}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Remove a role from a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Remove role from user
      tags:
      - Roles
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"